Extract `THRESHOLD_REGISTRY` and threshold configs into `thresholds.ts`
